requirements.txt
flask
gunicorn
supabase
requests
psycopg2-binary
paramiko
streamlit
pandas
openpyxl
python-dotenv
